Update to use plugin-collections instead of theme-mdx-collections
gatsby-theme-mdx-pages and mdx-collections ran out of time
on their fusion dance, and now they are completely separate entities.

...but actually, though, mdx-collections is better as a generalized
plugin instead of being specific to MdxPages, so that's the direction
it's going from now on. While it radically changes mdx-collections, this
theme only changes minimally.

The MdxPage getCollection is moved from the former theme to here, passed
into the now plugin much like plugin-taxonomies is handled.

gatsby-config and the readme are changed to reflect this as well.

Since the usage of the plugin hasn't changed despite the dependencies
being changed radically, only the minor version is bumped.